Refactor: Moved plotting functions outside of apps
